requirements.txt
APScheduler==3.6.3
boto3==1.16.25
botocore==1.19.25
certifi==2020.11.8
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
decorator==4.4.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
pycparser==2.20
python-dateutil==2.8.1
python-telegram-bot==13.0
pytz==2020.4
s3transfer==0.3.3
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
